layout and base classes for six primary screens
Each panel is a primary screen. Panel Test should give the just of the layout. It currently holds the SearchPanel information, which I need to tweak a little more then move to SearchPanel.java. The MainApplicationWindow is the overall layout (which is simply a menu bar with buttons on a frame. The buttons switch screens. We can work on the adjustability of the layout with screen resizing later when we tweak things, but for now stick with this layout size.