side windows
The uploaded files are a mixture of the setting and pop up windows as well as an updated MainPanel, MainApplicationWindow, and POSPanel. The POSPanel is a rough visual that will change with the backend code. The MainApplicationWindow includes the login window, but there is no backend set up for the login window. The MainPanel includes the change user and setting buttons in operation.