Add files via upload
NOTICE: 
DatabaseConnection and MainApplicationWindow have been modified. There is a new txt file called DatabaseName.txt and it is to be placed outside of the src. DatabaseName.txt holds the database name (pretty straight forward) and without it, the database will not connect. This is done for the future implemention process for other customers.